requirements.txt
boto3==1.35.36
python-dotenv==1.0.1
Pillow==10.4.0
imagehash==4.3.1
numpy==1.26.4
pandas==2.2.2
tqdm==4.66.5
chromadb==0.5.11
torch>=2.2.0
transformers>=4.44.0
accelerate>=0.33.0
sentencepiece>=0.2.0
sentence-transformers>=3.0.0
streamlit==1.38.0
